refactor: streamline file descriptor handling and improve syscall definitions

+++ Kernel/include/syscall_dispatcher.h
@@ -8,7 +8,7 @@
 #include "pipes.h"
 
 // syscalls de arqui
-static int      sys_write(uint64_t fd, const char *buf, uint64_t count);
+static int      sys_write(int fd, const char *buf, uint64_t count);
 static int      sys_read(int fd, char *buf, uint64_t count);
 static void     sys_date(uint8_t *buffer);
 static void     sys_time(uint8_t *buffer);
